refactor(Accordion): extracted inner components for the docgen to work properly
