+++ packages/picasso/src/Accordion/index.ts
@@ -4,3 +4,6 @@ import { Props } from './Accordion'
 
 export { default } from './Accordion'
 export type AccordionProps = OmitInternalProps<Props>
+
+export { default as Details } from './Details'
+export { default as Summary } from './Summary'
